requirements.txt
flask
google-genai
chromadb
PyMuPDF
uuid
flask-cors
python-dotenv